SBG_System/requirements.txt
# CrewAI & Orchestration
python-dotenv==1.0.0

# Deep Learning
torch>=2.2.0
tensorflow>=2.14.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.14.0

# Signal Processing
librosa>=0.10.0
pywavelets>=1.5.0

# ML Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# React/Frontend deps (installed via npm)
# Will be separate

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Report Generation
reportlab>=4.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
